fix push screen
