--- shared/push/push.js.flow
@@ -1,11 +1,0 @@
-// @flow
-import {Component} from 'react'
-
-import type {TokenType} from '../constants/push'
-
-export type Props = {
-  prompt: boolean,
-  permissionsRequesting: boolean,
-  configurePush: () => void,
-  onRequestPermissions: () => void,
-}
