+++ shared/settings/routes.native.js
@@ -6,7 +6,7 @@ import Settings from './'
 import InvitationsContainer from './invites/container'
 import InviteGenerated from './invite-generated'
 import Feedback from './feedback-container'
-import Push from '../push/push.native'
+import Push from '../app/push/push.native'
 import DevicesRoute from '../devices/routes'
 import WebLinks from './web-links'
